model regenerated

+++ lib/model/om/BaseNewsletterMailing.php
@@ -642,6 +642,11 @@ abstract class BaseNewsletterMailing extends BaseObject  implements Persistent
 			$deleteQuery = NewsletterMailingQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(NewsletterMailingPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "newsletter_mailings")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -685,6 +690,11 @@ abstract class BaseNewsletterMailing extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(NewsletterMailingPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "newsletter_mailings")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(NewsletterMailingPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -705,6 +715,11 @@ abstract class BaseNewsletterMailing extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(NewsletterMailingPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "newsletter_mailings")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(NewsletterMailingPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1494,6 +1509,26 @@ abstract class BaseNewsletterMailing extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(NewsletterMailingPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && NewsletterMailingPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return NewsletterMailingPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
