+++ lib/model/om/BaseSubscriberGroup.php
@@ -597,6 +597,11 @@ abstract class BaseSubscriberGroup extends BaseObject  implements Persistent
 			$deleteQuery = SubscriberGroupQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(SubscriberGroupPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "subscriber_groups")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -640,6 +645,11 @@ abstract class BaseSubscriberGroup extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(SubscriberGroupPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "subscriber_groups")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(SubscriberGroupPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -660,6 +670,11 @@ abstract class BaseSubscriberGroup extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(SubscriberGroupPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "subscriber_groups")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(SubscriberGroupPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1784,6 +1799,26 @@ abstract class BaseSubscriberGroup extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(SubscriberGroupPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && SubscriberGroupPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return SubscriberGroupPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
